clean: codegen snapshot_indexer - move 'generate_queries_without_timestamp' to utils
