+++ chainsight-cdk-macros/src/canisters/snapshot_indexer_https.rs
@@ -3,6 +3,8 @@ use proc_macro::TokenStream;
 use quote::{format_ident, quote};
 use syn::parse_macro_input;
 
+use crate::canisters::utils::generate_queries_without_timestamp;
+
 pub fn def_snapshot_indexer_https(input: TokenStream) -> TokenStream {
     let input_json_string: String = parse_macro_input!(input as syn::LitStr).value();
     let config: SnapshotIndexerHTTPSConfig = serde_json::from_str(&input_json_string).unwrap();
@@ -101,81 +103,6 @@ fn custom_code(config: SnapshotIndexerHTTPSConfig) -> proc_macro2::TokenStream {
         }
         #queries
         did_export!(#id);
-    }
-}
-
-pub fn generate_queries_without_timestamp(
-    return_type: proc_macro2::Ident,
-) -> proc_macro2::TokenStream {
-    let query_derives = quote! {
-        #[ic_cdk::query]
-        #[candid::candid_method(query)]
-    };
-    let update_derives = quote! {
-        #[ic_cdk::update]
-        #[candid::candid_method(update)]
-    };
-
-    quote! {
-        fn _get_last_snapshot_value() -> #return_type {
-            get_last_snapshot().value
-        }
-
-        fn _get_top_snapshot_values(n: u64) -> Vec<#return_type> {
-            get_top_snapshots(n).iter().map(|s| s.value.clone()).collect()
-        }
-
-        fn _get_snapshot_value(idx: u64) -> #return_type {
-            get_snapshot(idx).value
-        }
-
-        #query_derives
-        pub fn get_last_snapshot_value() -> #return_type {
-            _get_last_snapshot_value()
-        }
-
-        #query_derives
-        pub fn get_top_snapshot_values(n: u64) -> Vec<#return_type> {
-            _get_top_snapshot_values(n)
-        }
-
-        #query_derives
-        pub fn get_snapshot_value(idx: u64) -> #return_type {
-            _get_snapshot_value(idx)
-        }
-
-        #update_derives
-        pub async fn proxy_get_last_snapshot_value(input: std::vec::Vec<u8>) -> std::vec::Vec<u8> {
-            use chainsight_cdk::rpc::Receiver;
-            chainsight_cdk::rpc::ReceiverProviderWithoutArgs::<#return_type>::new(
-                proxy(),
-                _get_last_snapshot_value,
-            )
-            .reply(input)
-            .await
-        }
-
-        #update_derives
-        pub async fn proxy_get_top_snapshot_values(input: std::vec::Vec<u8>) -> std::vec::Vec<u8> {
-            use chainsight_cdk::rpc::Receiver;
-            chainsight_cdk::rpc::ReceiverProvider::<u64, Vec<#return_type>>::new(
-                proxy(),
-                _get_top_snapshot_values,
-            )
-            .reply(input)
-            .await
-        }
-
-        #update_derives
-        pub async fn proxy_get_snapshot_value(input: std::vec::Vec<u8>) -> std::vec::Vec<u8> {
-            use chainsight_cdk::rpc::Receiver;
-            chainsight_cdk::rpc::ReceiverProvider::<u64, #return_type>::new(
-                proxy(),
-                _get_snapshot_value,
-            )
-            .reply(input)
-            .await
-        }
     }
 }
