+++ chainsight-cdk-macros/src/canisters/snapshot_indexer_icp.rs
@@ -6,7 +6,7 @@ use proc_macro::TokenStream;
 use quote::{format_ident, quote};
 use syn::parse_macro_input;
 
-use super::snapshot_indexer_https::generate_queries_without_timestamp;
+use crate::canisters::utils::generate_queries_without_timestamp;
 
 pub fn def_snapshot_indexer_icp(input: TokenStream) -> TokenStream {
     let input_json_string: String = parse_macro_input!(input as syn::LitStr).value();
